// client/src/context/actions.tsx
export enum ActionTypes {
  DISPLAY_ALERT = "SHOW_ALERT",
  HIDE_ALERT = "HIDE_ALERT",

  REGISTER_USER_BEGIN = "REGISTER_USER_BEGIN",
  REGISTER_USER_SUCCESS = "REGISTER_USER_SUCCESS",
  REGISTER_USER_ERROR = "REGISTER_USER_ERROR",

  LOGIN_USER_BEGIN = "LOGIN_USER_BEGIN",
  LOGIN_USER_SUCCESS = "LOGIN_USER_SUCCESS",
  LOGIN_USER_ERROR = "LOGIN_USER_ERROR",

  LOGOUT_USER = "LOGOUT_USER",

  TOGGLE_SIDEBAR = "TOGGLE_SIDEBAR",
  SHOW_LOGOUT = "SHOW_LOGOUT",
}